Moved fuzzySearch to the string category
